requirements.txt
# requirements.txt for AI Text Generator project

python-dotenv
openai
